+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added parameter `--null-value` to produce command in order to produce a null value (tombstone)
+
 ## 1.21.0 - 2021-10-01
 
 ### Added